Move snapshot tests from ES to monitoring package
